Filter streets

Internal function for filtering streets by multiple parameters

Usage

filter_streets(
  x,
  sha_class = NULL,
  street_type = NULL,
  block_num = NULL,
  union = FALSE,
  bbox = NULL,
  call = caller_env()
)

Arguments

x

sf object with streets to filter

sha_class

selected SHA classifications to include. "all" selects all streets with an assigned SHA classification (around one-quarter of all street segments). Additional options include c("COLL", "LOC", "MART", "PART", "FWY", "INT")

street_type

selected street subtypes to include. By default, the returned data includes all subtypes except alleys ("STRALY"). Options include c("STRALY", "STRPRD", "STRR", "STREX", "STRFIC", "STRNDR", "STRURD", "STCLN", "STRTN"). Not supported for

block_num

Integer vector with block number, e.g. 300, or range of block numbers (e.g. c(100, 500)) to filter streets.

union

Logical. Default TRUE. Union geometry based on fullname of streets.

bbox

Bounding box to filter passed to location parameter of getdata::get_location_data().

call

The execution environment of a currently running function, e.g. caller_env(). The function will be mentioned in error messages as the source of the error. See the call argument of abort() for more information.

Value

streets filtered by parameters