Get data from an ArcGIS FeatureServer or MapServer

This function is deprecated because the functionality has been incorporated into the improved getdata::get_esri_data() function which uses a similar set of parameters.

Wraps the esri2sf::esri2sf() function to download an ArcGIS FeatureServer or MapServer.

Usage

get_area_esri_data(
  area = NULL,
  bbox = NULL,
  url = NULL,
  where = "1=1",
  type = c("md food stores 2017 2018", "farmers markets 2020",
    "baltimore food stores 2016", "baltimore demolitions", "contour 2ft",
    "contours 10ft", "open vacant building notices", "liquor licenses",
    "fixed speed cameras", "red light cameras", "edge of pavement"),
  dist = NULL,
  diag_ratio = NULL,
  asp = NULL,
  trim = FALSE,
  crs = pkgconfig::get_config("mapbaltimore.crs", 2804)
)

Arguments

area

sf object. Optional. Only used if trim is TRUE.

bbox

bbox object. Optional but suggested to avoid downloading entire layer. See sf::st_bbox() for more information.

url

FeatureServer or MapServer url to retrieve data from. Passed to url parameter of esri2sf::esri2sf() function.

where

string for where condition. Default is 1=1 for all rows.

type

Type of data to get. Options include "md food stores 2017 2018", "farmers markets 2020", "baltimore food stores 2016", "baltimore demolitions", "contour 2ft", "contours 10ft", "open vacant building notices", "liquor licenses", "fixed speed cameras", "red light cameras", and "edge of pavement"

dist

buffer distance in meters. Optional.

diag_ratio

ratio to set map extent based diagonal distance of area's bounding box. Ignored when dist is provided.

asp

Aspect ratio of width to height as a numeric value (e.g. 0.33) or character (e.g. "1:3").

trim

Logical. Default FALSE. If TRUE, area is required.

crs

Coordinate reference system. Default 2804.