Get real property data

Get showing parcels described as owner occupied, non-owner occupied, vacant, and unimproved. Real property or parcel data is from the Maryland State Department of Assessment and Taxation and may include errors.

Usage

get_area_property(
  area = NULL,
  bbox = NULL,
  dist = NULL,
  diag_ratio = NULL,
  unit = "m",
  asp = NULL,
  crop = TRUE,
  trim = FALSE,
  cache = FALSE,
  filename = NULL,
  overwrite = FALSE,
  ...
)

format_property_data(data)

Arguments

area

sf object. If multiple areas are provided, they are unioned into a single sf object using sf::st_union()

bbox

bbox object defining area used to filter data. If an area is provided, the bounding box is ignored.

dist

buffer distance in units. Optional.

diag_ratio

ratio of diagonal distance of area's bounding box used as buffer distance. e.g. if the diagonal distance is 3000 meters and the "diag_ratio = 0.1" a 300 meter will be used. Ignored when dist is provided.

unit

Units for buffer. Supported options include "meter", "foot", "kilometer", and "mile", "nautical mile" Common abbreviations (e.g. "km" instead of "kilometer") are also supported. Distance in units is converted to units matching GDAL units for x; defaults to "meter"

asp

Aspect ratio of width to height as a numeric value (e.g. 0.33) or character (e.g. "1:3"). If numeric, get_asp() returns the same value without modification.

crop

If TRUE, data cropped to area or bounding box sf::st_crop() adjusted by the dist, diag_ratio, and asp parameters provided. Default TRUE.

trim

If TRUE, data trimmed to area with sf::st_intersection(). This option is not supported for any adjusted areas that use the dist, diag_ratio, or asp parameters. Default FALSE.

cache

If TRUE, cache data to mapbaltimore cache folder. Defaults to FALSE.

filename

File name to use for cached file. Defaults to name of data. If the data is an sf object make sure to include the file type, e.g. "data.gpkg", supported by sf::write_sf(). All other data is written to rda with readr::write_rds().

overwrite

Logical. Default FALSE. If TRUE, overwrite any existing cached files that use the same filename.

...

Additional parameters passed to getdata::get_esri_data().

data

sf object including data in area

Examples

get_area_property(
  area = neighborhoods[1, ],
  dist = -150,
  unit = "m"
)
#> Error in get_area_property(area = neighborhoods[1, ], dist = -150, unit = "m"): 500
#> ℹ Service CityView/Realproperty/MapServer not found