Get selected area streets

Get streets within an area or areas.

Usage

get_area_streets(
  area = NULL,
  street_type = NULL,
  sha_class = NULL,
  bbox = NULL,
  dist = NULL,
  diag_ratio = NULL,
  asp = NULL,
  trim = FALSE,
  msa = FALSE,
  union = TRUE
)

Arguments

area

sf object with area of streets to return.

street_type

selected street subtypes to include. By default, the returned data includes all subtypes except alleys ("STRALY"). Options include c("STRALY", "STRPRD", "STRR", "STREX", "STRFIC", "STRNDR", "STRURD", "STCLN", "STRTN")

sha_class

selected SHA classifications to include. "all" selects all streets with an assigned SHA classification (around one-quarter of all street segments). Additional options include c("COLL", "LOC", "MART", "PART", "FWY", "INT")

bbox

bbox object defining area used to filter data. If an area is provided, the bounding box is ignored.

dist

buffer distance in meters. Optional.

diag_ratio

ratio to set map extent based diagonal distance of area's bounding box. Ignored when dist is provided.

asp

Aspect ratio of width to height as a numeric value (e.g. 0.33) or character (e.g. "1:3").

trim

Logical. Default FALSE. Trim streets to area using sf::st_intersection().

msa

Logical. Default FALSE. Get streets from cached baltimore_msa_streets.gpkg file using cachedata parameter of get_area_data function.

union

Logical. Default TRUE. Union geometry based on fullname of streets.