Get vacant building notices

Parcel boundaries for all properties with an active vacant building notice. If a building is unoccupied and unsafe or unfit for people to live or work inside the building, or has two code violations that have not been fixed, or has six code violations in the past year, then the building may receive a vacant building notice in Baltimore City.

Usage

get_area_vacants(
  area = NULL,
  bbox = NULL,
  dist = NULL,
  diag_ratio = NULL,
  asp = NULL,
  crop = TRUE,
  trim = FALSE,
  rehabbed = FALSE
)

Arguments

area

sf object. If multiple areas are provided, they are unioned into a single sf object using sf::st_union()

bbox

bbox object defining area used to filter data. If an area is provided, the bounding box is ignored.

dist

buffer distance in meters. Optional.

diag_ratio

ratio to set map extent based diagonal distance of area's bounding box. Ignored when dist is provided.

asp

Aspect ratio of width to height as a numeric value (e.g. 0.33) or character (e.g. "1:3").

crop

If TRUE, data cropped to area or bounding box sf::st_crop() adjusted by the dist, diag_ratio, and asp parameters provided. Default TRUE.

trim

If TRUE, data trimmed to area with sf::st_intersection(). This option is not supported for any adjusted areas that use the dist, diag_ratio, or asp parameters. Default FALSE.

rehabbed

If TRUE, return building permits pulled on properties with vacant building notices. Default FALSE.

Details

If the rehabbed parameter is TRUE, the returned data is use and occupancy permits that were pulled on properties with vacant building notices. DHCD uses this data as proxy for vacant building rehabs.