Get selected crime incidents from Open Baltimore data portal

This function uses the RSocrata package to download BPD Part 1 Victim-Based Crime Data from the Open Baltimore data portal. Crime data may be filtered by crime type, by date, or by neighborhood or police district. Unlike the get_citations and get_service_requests functions this function cannot filter by City Council district. Filtering by multiple criteria or limiting the period of time covered is recommended to avoid a long wait for a large data file to download.

get_crimes(
  crime_type = NULL,
  start_date = NULL,
  end_date = NULL,
  filter_by = c("neighborhood", "police district"),
  area_name = NULL,
  geometry = FALSE
)

Arguments

crime_type
Character vector matching one of the possible crime types.

start_date
The start date of the time period during which selected crimes occurred in the format "YYYY-MM-DD". An end date must be provided if a start date is provided.

end_date
The end date of the time period during which selected crimes occurred in the format "YYYY-MM-DD". An start date must be provided if an end date is provided.

filter_by
A character vector with the type of area to filter crimes by. Supported area types include neighborhoods ("neighborhood"), City Council districts ("council district"), and Police Districts ("police district").

area_name
A character vector with a the area name. Filtering by multiple areas or multiple area types is not currently supported.

geometry
If TRUE the selected crime data is converted to an sf object with a projected CRS (2804). Requests with no coordinates are excluded if TRUE. Defaults to FALSE.

Examples

if (FALSE) {
## Get all reported Part 1 crimes for January 2020 in the Southwest Police District.
get_crimes(
  start_date = "2020-01-01",
  end_date = "2020-01-31",
  filter_by = "police district",
  area_name = "SOUTHWEST"
)
}

if (FALSE) {
## Get all reported burglaries in 2019 in the Federal Hill neighborhood
get_crimes(
  crime_type = "BURGLARY",
  start_date = "2019-01-01",
  end_date = "2019-12-31",
  filter_by = "neighborhood",
  area_name = "FEDERAL HILL"
)
}