Get selected 311 Service Requests from Open Baltimore data portal

This function uses the RSocrata package to download 311 service requests from the Open Baltimore data portal. Requests may be one or more service request types, by date, or by neighborhood, City Council District, or Police District. Filtering by multiple criteria or for a limited period of time is recommended to avoid a long wait for a large data file to download.

get_service_requests(
  area,
  request_type = NULL,
  start_date = NULL,
  end_date = NULL,
  filter_by = c("neighborhood", "council_district", "police_district"),
  area_name = NULL,
  geometry = FALSE
)

Arguments

area
An sf object with a 'name' column. If area is provided, the function returns data within the bounding box of the provided area or areas.

request_type
A string or character vector with multiple strings matching one or more of the possible request_types.

start_date
The start date of the time period during which downloaded requests were created in the format "YYYY-MM-DD". An end date must be provided if a start date is provided.

end_date
The end date of the time period during which downloaded requests were created in the format "YYYY-MM-DD". An start date must be provided if an end date is provided.

filter_by
A character string of the type of area to filter requests by. Supported area types include neighborhoods ("neighborhood"), City Council districts ("council_district"), and Police Districts ("police_district").

area_name
A character string corresponding a neighborhood name or the number of a Council or Police District. Filtering by multiple areas or multiple area types is not currently supported.

geometry
If TRUE the service requests are converted to an sf object with a projected CRS (2804). Requests with no coordinates are excluded if TRUE. Defaults to FALSE.

Examples

if (FALSE) {
## Get service requests for leaf removal and overgrown trees/shrubs in Council District 12 from September 2019
get_service_requests(
  request_type = c("SW-Leaf Removal", "HCD-Trees and Shrubs"),
  start_date = "2019-09-01",
  end_date = "2019-09-30",
  filter_by = "council_district",
  area_name = "12"
)
}

if (FALSE) {
## Get service requests for snow and ice on sidewalks from January 2020
get_service_requests(
  request_type = "HCD-Snow and Ice on Sidewalks",
  start_date = "2020-01-01",
  end_date = "2020-01-31"
)
}