Get streets Get streets in Baltimore City by name with option to exclude streets by name, crop to a bounding box, or to filter to selected street types or functional classifications.

Get streets

Get streets in Baltimore City by name with option to exclude streets by name, crop to a bounding box, or to filter to selected street types or functional classifications.

get_streets(
  street_name = NULL,
  exclude_name = NULL,
  street_type = NULL,
  sha_class = NULL,
  block_num = NULL,
  bbox = NULL,
  union = TRUE
)

Arguments

street_name
Street names to return. Required.

exclude_name
Street names to exclude

street_type
selected street subtypes to include. By default, the returned data includes all subtypes except alleys ("STRALY"). Options include c("STRALY", "STRPRD", "STRR", "STREX", "STRFIC", "STRNDR", "STRURD", "STCLN", "STRTN"). Not supported for

sha_class
selected SHA classifications to include. "all" selects all streets with an assigned SHA classification (around one-quarter of all street segments). Additional options include c("COLL", "LOC", "MART", "PART", "FWY", "INT")

block_num
Integer vector with block number c(300) or range of block numbers (c(100, 500)) to filter streets.

bbox
bbox to crop returned streets. Optional.

union
Logical. If `TRUE`, use st_union to combine geometry by `fullname` of the streets.

Value

OUTPUT_DESCRIPTION

Details

DETAILS

See also

Examples

if (FALSE) {
get_streets(street_name = "UNIVERSITY PKWY")

get_streets(street_name = c("E FAYETTE", "ORLEANS"), block_num = c(1700, 3600))
}