Housing Market Typology 2017

The 2017 update of the City’s Housing Market Typology was jointly developed by the Baltimore City Planning Department, Department of Housing & Community Development, and The Reinvestment Fund.

Usage

hmt_2017

Format

A data frame with 663 rows and 15 variables:

geoid

U.S. Census Block Group GeoID

geoid_part

Identifier for U.S. Census Block Group GeoID including part identifier

part

Part identifier

cluster

Housing market cluster

cluster_group

Housing market cluster

median_sales_price

Median sales price, Q3 2015 - Q2 2017

sales_price_variation

Sales price variation, Q3 2015 - Q2 2017

num_sales

Number of residential sales, Q3 2015 - Q2 2017

num_foreclosure_filings

Number of foreclosure filings, Q3 2015 - Q2 2017

perc_foreclosure_sales

Percent of sales through foreclosure, Q3 2015 - Q2 2017

perc_homeowners

Percent owner occupied, July 2017

perc_permits_over10k

Percent of residential building permits over $10,000, Q3 2015 - Q2 2017

vacant_lots_bldgs_per_acre_res

Vacant lots and buildings per residential acre, July 2017

units_per_acre_res

Housing units per residential acre, July 2017

geometry

MULTIPOLYGON geometry matching Census blocks groups or parts of block groups

Source

https://opendata.baltimorecity.gov/egis/rest/services/Hosted/Housing_Market_Typology_2017/FeatureServer/0