Label area street names at selected locations

Label street names at selected locations.

label_area_streets(
  area,
  geom = c("label", "label_repel"),
  sha_class = NULL,
  label_location = c("area", "edge", "topright", "bottomleft")
)

Arguments

area
sf object. Labels returned for streets in or around area bounding box.

geom
Character vector matching name of geom returned with labels. "label" for "ggplot::geom_label" or "label_repel" for "ggrepel::geom_label_repel" are supported.

sha_class
Character vector. "all" selects all streets with an assigned SHA classification. Options include c("COLL", "LOC", "MART", "PART", "FWY", "INT")

label_location
Options include "area", "edge", "topright", or "bottomleft". Defaults to "area"