Baltimore City Real Property Data

This dataset represents the City of Baltimore parcel boundaries, with ownership, address, valation and other property information. This data was downloaded on October 6, 2020.

real_property

Format

A data frame with 238,340 rows and 88 variables:

objectid

...

pin

...

pinrelate

...

blocklot

...

block

...

lot

...

ward

...

section

...

assessor

...

taxbase

...

bfcvland

...

bfcvimpr

...

landexmp

...

imprexmp

...

citycred

...

statcred

...

ccredamt

...

scredamt

...

permhome

...

assesgrp

...

lot_size

...

no_imprv

...

currland

...

currimpr

...

exmpland

...

exmpimpr

...

fullcash

...

exmptype

...

exmpcode

...

usegroup

...

zonecode

...

sdatcode

...

artaxbas

...

distswch

...

dist_id

...

statetax

...

city_tax

...

ar_owner

...

deedbook

...

deedpage

...

saledate

...

owner_abbr

...

owner_1

...

owner_2

...

owner_3

...

fulladdr

...

stdirpre

...

st_name

...

st_type

...

bldg_no

...

fraction

...

unit_num

...

span_num

...

spanfrac

...

zip_code

...

extd_zip

...

dhcduse1

...

dhcduse2

...

dhcduse3

...

dhcduse4

...

dwelunit

...

eff_unit

...

roomunit

...

rpdeltag

...

respagcy

...

salepric

...

propdesc

...

neighbor

...

srvccntr

...

year_build

...

structarea

...

ldate

...

ownmde

...

grndrent

...

subtype_geodb

...

sdatlink

...

blockplat

...

mailtoadd

...

vacind

...

shape_st_area

...

shape_st_length

...

geometry

...

neighborhood

...

council_district

...

police_district

...

csa

...

block_group

...

tract

...

...

Source

https://gis-baltimore.opendata.arcgis.com/datasets/real-property