Scales for Baltimore data

Custom palettes for two package datasets: mta_bus_lines and hmt_2017 (both for cluster and cluster group).

Usage

scale_mapbaltimore(
  palette = NULL,
  values = NULL,
  na.value = "grey50",
  aesthetics = c("color", "fill"),
  error_call = rlang::caller_env(),
  ...
)

scale_color_mapbaltimore(palette = NULL, na.value = "grey50", ...)

scale_fill_mapbaltimore(palette = NULL, na.value = "grey50", ...)

Arguments

palette

Options include "mta_bus", "hmt_2017", "hmt_cluster", "cluster", "hmt_cluster_group", or "cluster_group", Default: NULL

values

a set of aesthetic values to map data values to. The values will be matched in order (usually alphabetical) with the limits of the scale, or with breaks if provided. If this is a named vector, then the values will be matched based on the names instead. Data values that don't match will be given na.value.

na.value

Defaults to "grey50"

aesthetics

Character string or vector of character strings listing the name(s) of the aesthetic(s) that this scale works with. This can be useful, for example, to apply colour settings to the colour and fill aesthetics at the same time, via aesthetics = c("colour", "fill").

...

Arguments passed on to discrete_scale

palette

A palette function that when called with a single integer argument (the number of levels in the scale) returns the values that they should take (e.g., scales::hue_pal()).

limits

One of:

NULL to use the default scale values

A character vector that defines possible values of the scale and their order

A function that accepts the existing (automatic) values and returns new ones. Also accepts rlang lambda function notation.

drop

Should unused factor levels be omitted from the scale? The default, TRUE, uses the levels that appear in the data; FALSE uses all the levels in the factor.

na.translate

Unlike continuous scales, discrete scales can easily show missing values, and do so by default. If you want to remove missing values from a discrete scale, specify na.translate = FALSE.

scale_name

The name of the scale that should be used for error messages associated with this scale.

name

The name of the scale. Used as the axis or legend title. If waiver(), the default, the name of the scale is taken from the first mapping used for that aesthetic. If NULL, the legend title will be omitted.

labels

One of:

NULL for no labels

waiver() for the default labels computed by the transformation object

A character vector giving labels (must be same length as breaks)

An expression vector (must be the same length as breaks). See ?plotmath for details.

A function that takes the breaks as input and returns labels as output. Also accepts rlang lambda function notation.

guide

A function used to create a guide or its name. See guides() for more information.

super

The super class to use for the constructed scale

Examples

if (FALSE) {
if (interactive()) {
  ggplot2::ggplot(data = dplyr::filter(mta_bus_lines, frequent)) +
    ggplot2::geom_sf(ggplot2::aes(color = route_abb), alpha = 0.5, size = 2) +
    scale_mapbaltimore(palette = "bus") +
    ggplot2::theme_minimal()

  ggplot2::ggplot(data = hmt_2017) +
    ggplot2::geom_sf(ggplot2::aes(fill = cluster_group, color = cluster_group)) +
    scale_mapbaltimore(palette = "cluster_group") +
    ggplot2::theme_minimal()
}
}