Neighborhood-to-U.S. Census Tract Crosswalk

Share of total households is based on the proportion of U.S. Census tract population within the named neighborhood based on overlapping U.S. Census Block groups.

Usage

xwalk_neighborhood2tract

Format

A data frame with 551 rows and 4 variables:

name

Neighborhood name

geoid

GeoID for U.S. Census tract

tract

Tract number

weight_households

Share of total households in neighborhood and U.S. Census tract (based on 2010 decennial Census). Variable code is "H013001".

weight_units

Share of occupied housing units in neighborhood and U.S. Census tract (based on 2020 decennial Census PL-94171 redistricting data). Variable code is "H1_002N".