Get area of selected administrative type

Get a sf object with one or more neighborhoods, Baltimore City Council districts, Maryland Legislative Districts, U.S. Congressional Districts, Baltimore Planning Districts, Baltimore Police Districts, or Community Statistical Areas.

get_area(
  type = c("neighborhood", "council district", "legislative district",
    "congressional district", "planning district", "police district", "csa"),
  area_name = NULL,
  area_id = NULL,
  union = FALSE
)

Arguments

type
area type matching one of the included boundary datasets. Supported values include c("neighborhood", "council district", "legislative district", "congressional district", "planning district", "police district", "csa")

area_name
name or names matching id column in data of selected dataset.

area_id
identifier or identifiers matching id column of selected dataset. Not all supported datasets have an id column

union
If TRUE and multiple area names are provided, the area geometry is combined with st_union and names are concatenated into a single string. Defaults to FALSE.

Examples

get_area(type = "neighborhood", area_name = "Harwood")

#> Simple feature collection with 1 feature and 1 field
#> geometry type:  MULTIPOLYGON
#> dimension:      XY
#> bbox:           xmin: 433352.4 ymin: 183379.4 xmax: 433677.9 ymax: 183987
#> projected CRS:  NAD83(HARN) / Maryland
#> # A tibble: 1 x 2
#>   name                                                                  geometry
#> * <chr>                                                       <MULTIPOLYGON [m]>
#> 1 Harwood (((433531.5 183984.9, 433484.8 183983.2, 433440 183980.4, 433420.8 18…

get_area(type = "council district", area_id = c(12, 14))

#> Simple feature collection with 2 features and 2 fields
#> geometry type:  MULTIPOLYGON
#> dimension:      XY
#> bbox:           xmin: 431075 ymin: 179871.4 xmax: 436683.8 ymax: 186408.6
#> projected CRS:  NAD83(HARN) / Maryland
#>   id                       geometry        name
#> 1 14 MULTIPOLYGON (((432649.3 18... District 14
#> 2 12 MULTIPOLYGON (((433676 1838... District 12

get_area(type = "planning district", area_id = c("East", "Southeast"), union = TRUE)

#> Simple feature collection with 2 features and 3 fields
#> geometry type:  MULTIPOLYGON
#> dimension:      XY
#> bbox:           xmin: 433455.8 ymin: 174664.5 xmax: 440590 ymax: 184534.6
#> projected CRS:  NAD83(HARN) / Maryland
#>          id                        name abb                       geometry
#> 1      East      East Planning District   E MULTIPOLYGON (((435402.8 18...
#> 2 Southeast Southeast Planning District  SE MULTIPOLYGON (((440563.2 18...