Get real property data

Get showing parcels described as owner occupied, non-owner occupied, vacant, and unimproved. Real property or parcel data is from the Maryland State Department of Assessment and Taxation and may include errors.

Usage

get_area_property(
  area = NULL,
  bbox = NULL,
  dist = NULL,
  diag_ratio = NULL,
  unit = "m",
  asp = NULL,
  crop = TRUE,
  trim = FALSE,
  cache = FALSE,
  filename = NULL,
  overwrite = FALSE,
  ...
)

format_property_data(data)

Arguments

area

sf object. If multiple areas are provided, they are unioned into a single sf object using sf::st_union()

bbox

bbox object defining area used to filter data. If an area is provided, the bounding box is ignored.

dist

buffer distance in units. Optional.

diag_ratio

ratio of diagonal distance of area's bounding box used as buffer distance. e.g. if the diagonal distance is 3000 meters and the "diag_ratio = 0.1" a 300 meter will be used. Ignored when dist is provided.

unit

Units for buffer. Supported options include "meter", "foot", "kilometer", and "mile", "nautical mile" Common abbreviations (e.g. "km" instead of "kilometer") are also supported. Distance in units is converted to units matching GDAL units for x; defaults to "meter"

asp

Aspect ratio of width to height as a numeric value (e.g. 0.33) or character (e.g. "1:3"). If numeric, get_asp() returns the same value without modification.

crop

If TRUE, data cropped to area or bounding box sf::st_crop() adjusted by the dist, diag_ratio, and asp parameters provided. Default TRUE.

trim

If TRUE, data trimmed to area with sf::st_intersection(). This option is not supported for any adjusted areas that use the dist, diag_ratio, or asp parameters. Default FALSE.

cache

If TRUE, cache data to mapbaltimore cache folder. Defaults to FALSE.

filename

File name to use for cached file. Defaults to name of data. If the data is an sf object make sure to include the file type, e.g. "data.gpkg", supported by sf::write_sf(). All other data is written to rda with readr::write_rds().

overwrite

Logical. Default FALSE. If TRUE, overwrite any existing cached files that use the same filename.

...

Additional parameters passed to getdata::get_esri_data().

data

sf object including data in area

Examples

get_area_property(
  area = neighborhoods[1, ],
  dist = -150,
  unit = "m"
)
#> ── Downloading "Property Information" from <https://geodata.baltimorecity.gov/eg
#> Layer type: "Feature Layer"
#> Geometry type: "esriGeometryPolygon"
#> Service CRS: "EPSG:2248"
#> Output CRS: "EPSG:3857"
#> 
#> Simple feature collection with 94 features and 86 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -8528406 ymin: 4768303 xmax: -8528315 ymax: 4768626
#> Projected CRS: WGS 84 / Pseudo-Mercator
#> First 10 features:
#>    objectid      pin pinrelate blocklot block lot ward section assessor taxbase
#> 1    126290 3857A001  3857A001 3857A001 3857A 001   12     200      331  199533
#> 2    126291 3857A002  3857A002 3857A002 3857A 002   12     200      331  183333
#> 3    126292 3857A003  3857A003 3857A003 3857A 003   12     200      331  193300
#> 4    126293 3857A004  3857A004 3857A004 3857A 004   12     200      331  183333
#> 5    126294 3857A005  3857A005 3857A005 3857A 005   12     200      331  183333
#> 6    126295 3857A006  3857A006 3857A006 3857A 006   12     200      331  186033
#> 7    126296 3857A007  3857A007 3857A007 3857A 007   12     200      331  185733
#> 8    126297 3857A008  3857A008 3857A008 3857A 008   12     200      331  186300
#> 9    126298 3857A009  3857A009 3857A009 3857A 009   12     200      331  204967
#> 10   126299 3857A010  3857A010 3857A010 3857A 010   12     200      331  311533
#>    bfcvland bfcvimpr landexmp imprexmp citycred statcred ccredamt scredamt
#> 1        NA   100200        0        0        0        0   764.58        0
#> 2        NA    86600        0        0        0        0    70.54        0
#> 3        NA    96900        0        0        0        0    44.17        0
#> 4        NA    86600        0        0        0        0   540.44        0
#> 5        NA    86600        0        0        0        0     0.00        0
#> 6        NA    88900        0        0        0        0    75.33        0
#> 7        NA    88600        0        0        0        0     0.00        0
#> 8        NA    89100        0        0        0        0     0.00        0
#> 9        NA   106700        0        0        0        0    68.15        0
#> 10       NA   196600        0        0        0        0   277.89        0
#>    permhome assesgrp lot_size no_imprv currland currimpr exmpland exmpimpr
#> 1         H        2   16X100        N    80000   129200        0        0
#> 2         H        2 15-6X100        N    80000   111700        0        0
#> 3         H        2 15-6X100        N    80000   121500        0        0
#> 4         H        2 15-6X100        N    80000   111700        0        0
#> 5         N        2 15-6X100        N    80000   111700        0        0
#> 6         H        2 15-6X100        N    80000   114600        0        0
#> 7         N        2 15-6X100        N    80000   114300        0        0
#> 8         H        2 15-6X100        N    80000   114900        0        0
#> 9         H        2 15-6X100        N    80000   134100        0        0
#> 10        H        2 15-6X100        N    80000   249000        0        0
#>    fullcash exmptype exmpcode usegroup zoning sdatcode artaxbas distswch
#> 1         0     <NA>       00        R    R-7    11130   199533        Y
#> 2         0     <NA>       00        R    R-7    11130   183333        Y
#> 3         0     <NA>       00        R    R-7    11130   193300        Y
#> 4         0     <NA>       00        R    R-7    11130   183333        Y
#> 5         0     <NA>       00        R    R-7    11130   183333        Y
#> 6         0     <NA>       00        R    R-7    11130   186033        Y
#> 7         0     <NA>       00        R    R-7    11130   185733        Y
#> 8         0     <NA>       00        R    R-7    11130   186300        Y
#> 9         0     <NA>       00        R    R-7    11130   204967        Y
#> 10        0     <NA>       00        R    R-7    11130   311533        Y
#>    dist_id statetax city_tax ar_owner deedbook deedpage sale_date owner_abbr
#> 1        3   223.48  4485.50        H PMB07553     0345  07301998         NA
#> 2        3   205.33  4121.33        H FMC12730     0091  06232010         NA
#> 3        3   216.50  4345.38        H XAC27165     0252  08292024         NA
#> 4        3   205.33  4121.33        H FMC08531     0007  06141999         NA
#> 5        3   205.33  4121.33        N MB 23382     0004  07132021         NA
#> 6        3   208.36  4182.02        H FMC13489     0411  05102011         NA
#> 7        3   208.02  4175.28        N FMC01494     0462  06272001         NA
#> 8        3   208.66  4188.02        H MB 25364     0420  12092022         NA
#> 9        3   229.56  4607.66        H MB 19079     0214  04242017         NA
#> 10       3   348.92  7003.26        H MB 24554     0081  05022022         NA
#>                 owner_1          owner_2 owner_3   full_address
#> 1    WEINSTEIN, MATTHEW KESSLER, LAURA T    <NA> 3001 ABELL AVE
#> 2        WEBER, JESSICA             <NA>    <NA> 3003 ABELL AVE
#> 3    HALL, LAURA COOPER             <NA>    <NA> 3005 ABELL AVE
#> 4        MOORE, BETTY A             <NA>    <NA> 3007 ABELL AVE
#> 5         WESOLVE54 LLC             <NA>    <NA> 3009 ABELL AVE
#> 6        WALSH, EAMON F             <NA>    <NA> 3011 ABELL AVE
#> 7  JACKSON, MARGARET A.             <NA>    <NA> 3013 ABELL AVE
#> 8           BERGER, MAX DODSON, KIMBERLY    <NA> 3015 ABELL AVE
#> 9          ADAMS, KEITH     ADAMS, NAOMI    <NA> 3017 ABELL AVE
#> 10       EMBRY, JULIA M             <NA>    <NA> 3019 ABELL AVE
#>    street_dir_prefix street_name street_type bldg_num fraction unit_num
#> 1               <NA>       ABELL         AVE     3001     <NA>     <NA>
#> 2               <NA>       ABELL         AVE     3003     <NA>     <NA>
#> 3               <NA>       ABELL         AVE     3005     <NA>     <NA>
#> 4               <NA>       ABELL         AVE     3007     <NA>     <NA>
#> 5               <NA>       ABELL         AVE     3009     <NA>     <NA>
#> 6               <NA>       ABELL         AVE     3011     <NA>     <NA>
#> 7               <NA>       ABELL         AVE     3013     <NA>     <NA>
#> 8               <NA>       ABELL         AVE     3015     <NA>     <NA>
#> 9               <NA>       ABELL         AVE     3017     <NA>     <NA>
#> 10              <NA>       ABELL         AVE     3019     <NA>     <NA>
#>    span_num spanfrac zip_code zip_code_ext dhcd_use dhcduse2 dhcduse3 dhcduse4
#> 1         0     <NA>    21218         3913     1008     0000     0000     0000
#> 2         0     <NA>    21218         3913     1008     0000     0000     0000
#> 3         0     <NA>    21218         3913     1008     0000     0000     0000
#> 4         0     <NA>    21218         3913     1008     0000     0000     0000
#> 5         0     <NA>    21218         3913     1008     0000     0000     0000
#> 6         0     <NA>    21218         3913     1008     0000     0000     0000
#> 7         0     <NA>    21218         3913     1008     0000     0000     0000
#> 8         0     <NA>    21218         3913     1008     0000     0000     0000
#> 9         0     <NA>    21218         3913     1008     0000     0000     0000
#> 10        0     <NA>    21218         3913     1008     0000     0000     0000
#>    dwelunit eff_unit roomunit rpdeltag agency sale_price propdesc neighborhood
#> 1         1        0        0     <NA>   <NA>      79000     <NA>        ABELL
#> 2         1        0        0     <NA>   <NA>     215000     <NA>        ABELL
#> 3         1        0        0     <NA>   <NA>     380000     <NA>        ABELL
#> 4         1        0        0     <NA>   <NA>      62000     <NA>        ABELL
#> 5         1        0        0     <NA>   <NA>      50000     <NA>        ABELL
#> 6         1        0        0     <NA>   <NA>     135000     <NA>        ABELL
#> 7         1        0        0     <NA>   <NA>          0     <NA>        ABELL
#> 8         1        0        0     <NA>   <NA>     250000     <NA>        ABELL
#> 9         1        0        0     <NA>   <NA>     260000     <NA>        ABELL
#> 10        1        0        0     <NA>   <NA>     325000     <NA>        ABELL
#>    srvccntr year_built structarea    ldate ownmde grndrent subtype_geodb
#> 1         5       1900       1900 11242024      F        0             1
#> 2         5       1900       1654 11242024      F        0             1
#> 3         5       1900       1654 11242024      F        0             1
#> 4         5       1900       1654 11242024      L       90             1
#> 5         5       1900       1654 11242024      F        0             1
#> 6         5       1900       1654 11242024      L       90             1
#> 7         5       1900       1654 11242024      L       90             1
#> 8         5       1900       1654 11242024      F        0             1
#> 9         5       1900       1654 11242024      L       78             1
#> 10        5       1900       1654 11242024      F        0             1
#>                                                                                                                             sdatlink
#> 1  http://sdat.dat.maryland.gov/realproperty/pages/viewdetails.aspx?County=03&SearchType=ACCT&Ward=12&SECTION=20&BLOCK=3857A&LOT=001
#> 2  http://sdat.dat.maryland.gov/realproperty/pages/viewdetails.aspx?County=03&SearchType=ACCT&Ward=12&SECTION=20&BLOCK=3857A&LOT=002
#> 3  http://sdat.dat.maryland.gov/realproperty/pages/viewdetails.aspx?County=03&SearchType=ACCT&Ward=12&SECTION=20&BLOCK=3857A&LOT=003
#> 4  http://sdat.dat.maryland.gov/realproperty/pages/viewdetails.aspx?County=03&SearchType=ACCT&Ward=12&SECTION=20&BLOCK=3857A&LOT=004
#> 5  http://sdat.dat.maryland.gov/realproperty/pages/viewdetails.aspx?County=03&SearchType=ACCT&Ward=12&SECTION=20&BLOCK=3857A&LOT=005
#> 6  http://sdat.dat.maryland.gov/realproperty/pages/viewdetails.aspx?County=03&SearchType=ACCT&Ward=12&SECTION=20&BLOCK=3857A&LOT=006
#> 7  http://sdat.dat.maryland.gov/realproperty/pages/viewdetails.aspx?County=03&SearchType=ACCT&Ward=12&SECTION=20&BLOCK=3857A&LOT=007
#> 8  http://sdat.dat.maryland.gov/realproperty/pages/viewdetails.aspx?County=03&SearchType=ACCT&Ward=12&SECTION=20&BLOCK=3857A&LOT=008
#> 9  http://sdat.dat.maryland.gov/realproperty/pages/viewdetails.aspx?County=03&SearchType=ACCT&Ward=12&SECTION=20&BLOCK=3857A&LOT=009
#> 10 http://sdat.dat.maryland.gov/realproperty/pages/viewdetails.aspx?County=03&SearchType=ACCT&Ward=12&SECTION=20&BLOCK=3857A&LOT=010
#>                                                    blockplat
#> 1  https://gis.baltimorecity.gov/zoning/blockplats/3857A.pdf
#> 2  https://gis.baltimorecity.gov/zoning/blockplats/3857A.pdf
#> 3  https://gis.baltimorecity.gov/zoning/blockplats/3857A.pdf
#> 4  https://gis.baltimorecity.gov/zoning/blockplats/3857A.pdf
#> 5  https://gis.baltimorecity.gov/zoning/blockplats/3857A.pdf
#> 6  https://gis.baltimorecity.gov/zoning/blockplats/3857A.pdf
#> 7  https://gis.baltimorecity.gov/zoning/blockplats/3857A.pdf
#> 8  https://gis.baltimorecity.gov/zoning/blockplats/3857A.pdf
#> 9  https://gis.baltimorecity.gov/zoning/blockplats/3857A.pdf
#> 10 https://gis.baltimorecity.gov/zoning/blockplats/3857A.pdf
#>                                    mailtoadd vacind shape_st_area
#> 1                      3001 ABELL AVE, 21218      N      1509.458
#> 2                      3003 ABELL AVE, 21218      N      1509.907
#> 3                      3005 ABELL AVE, 21218      N      1634.483
#> 4                      3007 ABELL AVE, 21218      N      1567.804
#> 5            8221 PLEASANT PLAINTS RD, 21286      N      1557.152
#> 6   1560 FALLING BROOK CT ODENTON, MD, 21113      N      1481.055
#> 7  950 SHOREPOINT CT #202 ALAMEDA, CA, 94501      N      1561.153
#> 8                      3015 ABELL AVE, 21218      N      1517.061
#> 9                      3017 ABELL AVE, 21218      N      1631.357
#> 10                     3019 ABELL AVE, 21218      N      1549.852
#>    shape_st_length block_num bldg_num_even_odd block_number_st vacant_lot
#> 1         231.0690      3000               Odd 3000  ABELL AVE      FALSE
#> 2         231.0577      3000               Odd 3000  ABELL AVE      FALSE
#> 3         233.5516      3000               Odd 3000  ABELL AVE      FALSE
#> 4         232.2099      3000               Odd 3000  ABELL AVE      FALSE
#> 5         232.0083      3000               Odd 3000  ABELL AVE      FALSE
#> 6         230.4667      3000               Odd 3000  ABELL AVE      FALSE
#> 7         232.0692      3000               Odd 3000  ABELL AVE      FALSE
#> 8         231.1979      3000               Odd 3000  ABELL AVE      FALSE
#> 9         233.4605      3000               Odd 3000  ABELL AVE      FALSE
#> 10        231.8600      3000               Odd 3000  ABELL AVE      FALSE
#>    vacant_bldg                       geometry
#> 1        FALSE MULTIPOLYGON (((-8528392 47...
#> 2        FALSE MULTIPOLYGON (((-8528352 47...
#> 3        FALSE MULTIPOLYGON (((-8528353 47...
#> 4        FALSE MULTIPOLYGON (((-8528353 47...
#> 5        FALSE MULTIPOLYGON (((-8528353 47...
#> 6        FALSE MULTIPOLYGON (((-8528353 47...
#> 7        FALSE MULTIPOLYGON (((-8528354 47...
#> 8        FALSE MULTIPOLYGON (((-8528354 47...
#> 9        FALSE MULTIPOLYGON (((-8528354 47...
#> 10       FALSE MULTIPOLYGON (((-8528354 47...