Get area 311 service requests from Open Baltimore

Get 311 service requests for a specific area. Service requests from 2017 to 2020 area available but only a single year can be requested at a time. Duplicate requests are removed from the returned data. Requests can be filtered by request type, responsible city agency, or both. You can return multiple types or agencies, by using a custom where query parameter or by calling each type/agency separately.

Usage

get_area_requests(
  area,
  year = 2022,
  date_range = NULL,
  request_type = NULL,
  agency = NULL,
  where = "1=1",
  dist = NULL,
  diag_ratio = NULL,
  unit = "m",
  asp = NULL,
  trim = FALSE,
  geometry = TRUE,
  crs = pkgconfig::get_config("mapbaltimore.crs", 2804),
  duplicates = FALSE
)

Arguments

year

Year for service requests. Default 2021. 2017 to 2022 supported.

request_type

Service request type.

agency

City agency responsible for request. Options include "Transportation", "BGE", "Solid Waste", "Housing", "Water Wastewater", "Health", "Call Center", "Finance", "Liquor Board", "Recreation & Parks", "Fire Department", "Parking Authority", and "General Services"

where

string for where condition. This parameter is ignored if a request_type or agency are provided.

dist

buffer distance in units. Optional.

diag_ratio

ratio of diagonal distance of area's bounding box used as buffer distance. e.g. if the diagonal distance is 3000 meters and the "diag_ratio = 0.1" a 300 meter will be used. Ignored when dist is provided.

unit

Units for buffer. Supported options include "meter", "foot", "kilometer", and "mile", "nautical mile" Common abbreviations (e.g. "km" instead of "kilometer") are also supported. Distance in units is converted to units matching GDAL units for x; defaults to "meter"

asp

Aspect ratio of width to height as a numeric value (e.g. 0.33) or character (e.g. "1:3"). If numeric, get_asp() returns the same value without modification.

geometry

Default TRUE. If FALSE, return requests with missing latitude/longitude (for years prior to 2021 only).

crs

Cordinate reference system to return, Default: 4326 for sf_to_df() and NULL for df_to_sf().

duplicates

If TRUE, return 311 service requests marked as "Duplicate". If FALSE, filter duplicate requests out of results.

Examples

# Get boundary for Edmondson Village
area <- get_area("neighborhood", "Edmondson Village")

# Get fallen limb requests for 2022
get_area_requests(
  area = area,
  date_range = c("2022-11-01", "2022-12-31"),
  request_type = "FOR-Fallen Limb"
)
#> ── Downloading "Customer_Service_Request311_2021_Present" from <https://egis.bal
#> Layer type: "Feature Layer"
#> Geometry type: "esriGeometryPoint"
#> Service CRS: "EPSG:2248"
#> Output CRS: "EPSG:2804"
#> 
#> Simple feature collection with 3 features and 21 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 426988.6 ymin: 181491.5 xmax: 427213.4 ymax: 181615.3
#> Projected CRS: NAD83(HARN) / Maryland
#>   service_request_num         sr_type method_received        created_date
#> 1         22-00893386 FOR-Fallen Limb           Phone 2022-11-11 13:16:36
#> 2         22-00926488 FOR-Fallen Limb           Phone 2022-11-27 09:20:02
#> 3         22-00990947 FOR-Fallen Limb           Phone 2022-12-23 06:36:22
#>   sr_status                                                  sr_status_url
#> 1    Closed https://balt311.baltimorecity.gov/citizen/requests/22-00893386
#> 2    Closed https://balt311.baltimorecity.gov/citizen/requests/22-00926488
#> 3      Open https://balt311.baltimorecity.gov/citizen/requests/22-00990947
#>           status_date            due_date          close_date
#> 1 2022-11-14 04:36:55 2022-11-14 13:16:36 2022-11-14 04:36:55
#> 2 2023-01-11 01:04:40 2022-11-28 09:20:02 2023-01-11 01:04:40
#> 3 2022-12-23 06:36:21 2022-12-26 06:36:21                <NA>
#>                                               agency    last_activity
#> 1 Recreation & Parks                                 Service Response
#> 2 Recreation & Parks                                 Service Response
#> 3 Recreation & Parks                                 Service Response
#>    last_activity_date                                             outcome
#> 1 2022-11-14 04:36:55 SR closed by agency-specific work management system
#> 2 2023-01-11 01:04:40 SR closed by agency-specific work management system
#> 3                <NA>                                                <NA>
#>   days_to_close              address zip_code      neighborhood
#> 1          2.64 1303 N WOODINGTON RD    21229 Edmondson Village
#> 2         44.66 1309 N WOODINGTON RD    21229 Edmondson Village
#> 3            NA    4114 MOUNTWOOD RD    21229 Edmondson Village
#>   council_district police_district latitude longitude                  geometry
#> 1                8    Southwestern 39.30233 -76.68711 POINT (426988.6 181615.3)
#> 2                8    Southwestern 39.30230 -76.68657 POINT (427034.9 181612.2)
#> 3                8    Southwestern 39.30121 -76.68450 POINT (427213.4 181491.5)

# Get dirty alley service requests for multiple years using purrr::map_dfr()
purrr::map_dfr(
  c(2021, 2020),
  ~ get_area_requests(
    area = area,
    year = .x,
    request_type = "SW-Dirty Alley"
  )
)
#> ── Downloading "Customer_Service_Request311_2021_Present" from <https://egis.bal
#> Layer type: "Feature Layer"
#> Geometry type: "esriGeometryPoint"
#> Service CRS: "EPSG:2248"
#> Output CRS: "EPSG:2804"
#> 
#> Removing 1 duplicate 311 service request.
#> ── Downloading "311_Customer_Service_Requests_2020.csv" from <https://opendata.b
#> Layer type: "Table"
#> 
#> Removing 5 duplicate 311 service requests.
#> Simple feature collection with 119 features and 22 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 426947.9 ymin: 180681.6 xmax: 427620.5 ymax: 181472.1
#> Projected CRS: NAD83(HARN) / Maryland
#> First 10 features:
#>    service_request_num        sr_type method_received        created_date
#> 1          21-00053788 SW-Dirty Alley           Phone 2021-01-22 03:07:54
#> 2          21-00059271 SW-Dirty Alley           Phone 2021-01-24 11:54:01
#> 3          21-00051584 SW-Dirty Alley             API 2021-01-21 08:36:09
#> 4          21-00107202 SW-Dirty Alley           Phone 2021-02-12 03:57:40
#> 5          21-00110903 SW-Dirty Alley           Phone 2021-02-13 05:44:24
#> 6          21-00115345 SW-Dirty Alley             API 2021-02-15 06:46:25
#> 7          21-00142040 SW-Dirty Alley           Phone 2021-02-24 09:37:26
#> 8          21-00062237 SW-Dirty Alley           Phone 2021-01-25 10:55:46
#> 9          21-00139354 SW-Dirty Alley             API 2021-02-23 17:06:52
#> 10         21-00070593 SW-Dirty Alley          System 2021-01-28 08:57:46
#>    sr_status                                                  sr_status_url
#> 1     Closed https://balt311.baltimorecity.gov/citizen/requests/21-00053788
#> 2     Closed https://balt311.baltimorecity.gov/citizen/requests/21-00059271
#> 3     Closed https://balt311.baltimorecity.gov/citizen/requests/21-00051584
#> 4     Closed https://balt311.baltimorecity.gov/citizen/requests/21-00107202
#> 5     Closed https://balt311.baltimorecity.gov/citizen/requests/21-00110903
#> 6     Closed https://balt311.baltimorecity.gov/citizen/requests/21-00115345
#> 7     Closed https://balt311.baltimorecity.gov/citizen/requests/21-00142040
#> 8     Closed https://balt311.baltimorecity.gov/citizen/requests/21-00062237
#> 9     Closed https://balt311.baltimorecity.gov/citizen/requests/21-00139354
#> 10    Closed https://balt311.baltimorecity.gov/citizen/requests/21-00070593
#>            status_date            due_date          close_date
#> 1  2021-01-29 10:01:00 2021-02-02 03:07:54 2021-01-29 10:01:21
#> 2  2021-02-03 09:12:00 2021-02-02 11:54:01 2021-02-03 09:12:54
#> 3  2021-01-22 09:34:00 2021-02-01 08:36:09 2021-01-22 09:34:27
#> 4  2021-02-23 09:24:00 2021-02-24 03:57:40 2021-02-23 09:25:02
#> 5  2021-02-16 10:24:00 2021-02-24 05:44:23 2021-02-16 10:24:58
#> 6  2021-02-23 09:26:00 2021-02-24 06:46:25 2021-02-23 09:27:20
#> 7  2021-03-04 10:18:00 2021-03-05 09:37:26 2021-03-04 10:18:37
#> 8  2021-01-25 10:55:00 2021-02-03 10:55:46 2021-01-25 10:56:15
#> 9  2021-03-03 09:21:00 2021-03-04 17:06:52 2021-03-03 09:21:41
#> 10 2021-02-04 10:51:00 2021-02-08 08:57:46 2021-02-04 10:51:52
#>                                                agency    last_activity
#> 1  Solid Waste                                        Service Response
#> 2  Solid Waste                                        Service Response
#> 3  Solid Waste                                        Service Response
#> 4  Solid Waste                                        Service Response
#> 5  Solid Waste                                        Service Response
#> 6  Solid Waste                                        Service Response
#> 7  Solid Waste                                        Service Response
#> 8  Solid Waste                                        Service Response
#> 9  Solid Waste                                        Service Response
#> 10 Solid Waste                                        Service Response
#>     last_activity_date                                        outcome
#> 1  2021-01-29 10:01:00                                 Work completed
#> 2  2021-02-03 09:12:00                                 Work completed
#> 3  2021-01-22 09:34:00                                 Work completed
#> 4  2021-02-23 09:24:00 SR assessed and no cause for action determined
#> 5  2021-02-16 10:24:00 SR assessed and no cause for action determined
#> 6  2021-02-23 09:26:00 SR assessed and no cause for action determined
#> 7  2021-03-04 10:18:00                                 Work completed
#> 8  2021-01-25 10:55:00                                 Work completed
#> 9  2021-03-03 09:21:00 SR assessed and no cause for action determined
#> 10 2021-02-04 10:51:00                                 Work completed
#>    days_to_close             address zip_code      neighborhood
#> 1           7.29   800 WILDWOOD PKWY    21229 Edmondson Village
#> 2           9.89 826 N WOODINGTON RD    21229    Rognel Heights
#> 3           1.04  905 MOUNT HOLLY ST    21229          Edgewood
#> 4          11.23    4000 COLBORNE RD    21229 Edmondson Village
#> 5           3.19      3801 ROKEBY RD    21229 Edmondson Village
#> 6           8.11    4009 COLBORNE RD    21229 Edmondson Village
#> 7           8.03 631 N WOODINGTON RD    21229 Edmondson Village
#> 8           0.00  704 MOUNT HOLLY ST    21229 Edmondson Village
#> 9           7.68     4010 GELSTON DR    21229 Edmondson Village
#> 10          7.08    3909 COLBORNE RD    21229 Edmondson Village
#>    council_district police_district latitude longitude objectid
#> 1                 8    Southwestern 39.29641 -76.68416       NA
#> 2                 8    Southwestern 39.29722 -76.68637       NA
#> 3                 8    Southwestern 39.29815 -76.67980       NA
#> 4                 8    Southwestern 39.29801 -76.68264       NA
#> 5                 8    Southwestern 39.29923 -76.68060       NA
#> 6                 8    Southwestern 39.29766 -76.68282       NA
#> 7                 8    Southwestern 39.29505 -76.68574       NA
#> 8                 8    Southwestern 39.29535 -76.67989       NA
#> 9                 8    Southwestern 39.29579 -76.68288       NA
#> 10                8    Southwestern 39.29768 -76.68213       NA
#>                     geometry
#> 1  POINT (427245.1 180958.8)
#> 2  POINT (427053.6 181048.6)
#> 3  POINT (427619.9 181153.5)
#> 4  POINT (427375.7 181136.8)
#> 5  POINT (427550.9 181273.1)
#> 6  POINT (427359.6 181098.7)
#> 7  POINT (427109.2 180807.5)
#> 8  POINT (427613.6 180842.2)
#> 9  POINT (427355.8 180890.3)
#> 10 POINT (427419.8 181100.8)