Get area 311 service requests from Open Baltimore

Get 311 service requests for a specific area. Service requests from 2017 to 2020 area available but only a single year can be requested at a time. Duplicate requests are removed from the returned data. Requests can be filtered by request type, responsible city agency, or both. You can return multiple types or agencies, by using a custom where query parameter or by calling each type/agency separately.

Usage

get_area_requests(
  area,
  year = 2021,
  request_type = NULL,
  agency = NULL,
  where = "1=1",
  dist = NULL,
  diag_ratio = NULL,
  asp = NULL,
  trim = FALSE,
  geometry = TRUE,
  crs = pkgconfig::get_config("mapbaltimore.crs", 2804),
  duplicates = FALSE
)

Arguments

area

sf object. Optional. Only used if trim is TRUE.

year

Year for service requests. Default 2021. 2017 to 2022 supported.

request_type

Service request type.

agency

City agency responsible for request. Options include "Transportation", "BGE", "Solid Waste", "Housing", "Water Wastewater", "Health", "Call Center", "Finance", "Liquor Board", "Recreation & Parks", "Fire Department", "Parking Authority", and "General Services"

where

string for where condition. This parameter is ignored if a request_type or agency are provided.

dist

buffer distance in meters. Optional.

diag_ratio

ratio to set map extent based diagonal distance of area's bounding box. Ignored when dist is provided.

asp

Aspect ratio of width to height as a numeric value (e.g. 0.33) or character (e.g. "1:3").

trim

Logical. Default FALSE. If TRUE, area is required.

geometry

Default TRUE. If FALSE, return requests with missing latitude/longitude (for years prior to 2021 only).

crs

Coordinate reference system. Default 2804.

duplicates

If TRUE, return 311 service requests marked as "Duplicate". If FALSE, filter duplicate requests out of results.

Examples

# Get boundary for Edmondson Village
area <- get_area("neighborhood", "Edmondson Village")

# Get fallen limb requests for 2022
get_area_requests(
  area = area,
  year = 2022,
  request_type = "FOR-Fallen Limb"
)
#> ✔ Downloading "Customer_Service_Request311_2021_Present" from
#>   <https://egis.baltimorecity.gov/egis/rest/services/GeoSpatialized_Tables/ServiceRequest_311/FeatureServer/0>
#> Layer type: "Feature Layer"
#> Geometry type: "esriGeometryPoint"
#> Service CRS: "EPSG:2248"
#> Output CRS: "EPSG:2804"
#> Removing 2 duplicate 311 service requests.
#> Simple feature collection with 26 features and 20 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 426977.6 ymin: 180708 xmax: 427576.8 ymax: 181615.3
#> Projected CRS: NAD83(HARN) / Maryland
#> First 10 features:
#>    service_request_num         sr_type method_received        created_date
#> 1          21-00247868 FOR-Fallen Limb           Phone 2021-04-05 09:43:19
#> 2          21-00461356 FOR-Fallen Limb           Phone 2021-06-18 15:49:24
#> 3          21-00497654 FOR-Fallen Limb             API 2021-07-02 05:35:53
#> 4          21-00504423 FOR-Fallen Limb             API 2021-07-05 18:50:25
#> 5          21-00510918 FOR-Fallen Limb           Phone 2021-07-07 13:02:59
#> 6          21-00510953 FOR-Fallen Limb           Phone 2021-07-07 13:12:35
#> 8          21-00513298 FOR-Fallen Limb           Phone 2021-07-08 09:00:52
#> 9          21-00528043 FOR-Fallen Limb           Phone 2021-07-13 12:18:58
#> 10         21-00616981 FOR-Fallen Limb           Phone 2021-08-14 03:34:27
#> 12         21-00617198 FOR-Fallen Limb           Phone 2021-08-14 05:08:28
#>    sr_status         status_date            due_date          close_date
#> 1     Closed 2021-04-08 04:59:32 2021-04-06 09:43:19 2021-04-08 04:59:33
#> 2     Closed 2021-06-22 05:52:43 2021-06-21 15:49:24 2021-06-22 05:52:44
#> 3     Closed 2021-07-02 05:45:19 2021-07-06 05:35:53 2021-07-02 05:45:20
#> 4     Closed 2021-07-09 05:30:20 2021-07-06 18:50:25 2021-07-09 05:30:21
#> 5     Closed 2021-07-12 08:35:26 2021-07-08 13:02:58 2021-07-12 08:35:27
#> 6     Closed 2021-07-12 11:56:54 2021-07-08 13:12:35 2021-07-12 11:56:56
#> 8     Closed 2021-07-12 08:33:50 2021-07-09 09:00:52 2021-07-12 08:33:51
#> 9     Closed 2021-07-14 10:06:59 2021-07-14 12:18:58 2021-07-14 10:07:00
#> 10    Closed 2021-08-14 07:21:47 2021-08-16 03:34:27 2021-08-14 07:21:48
#> 12    Closed 2021-08-17 12:29:50 2021-08-16 05:08:28 2021-08-17 12:29:51
#>                agency    last_activity  last_activity_date
#> 1  Recreation & Parks Service Response 2021-04-08 04:59:32
#> 2  Recreation & Parks Service Response 2021-06-22 05:52:43
#> 3  Recreation & Parks Service Response 2021-07-02 05:45:19
#> 4  Recreation & Parks Service Response 2021-07-09 05:30:20
#> 5  Recreation & Parks Service Response 2021-07-12 08:35:26
#> 6  Recreation & Parks Service Response 2021-07-12 11:56:54
#> 8  Recreation & Parks Service Response 2021-07-12 08:33:50
#> 9  Recreation & Parks Service Response 2021-07-14 10:06:59
#> 10 Recreation & Parks Service Response 2021-08-14 07:21:47
#> 12 Recreation & Parks Service Response 2021-08-17 12:29:50
#>                                                outcome days_to_close
#> 1  SR closed by agency-specific work management system          2.80
#> 2  SR closed by agency-specific work management system          3.59
#> 3  SR closed by agency-specific work management system          0.01
#> 4  SR closed by agency-specific work management system          3.44
#> 5  SR closed by agency-specific work management system          4.81
#> 6  SR closed by agency-specific work management system          4.95
#> 8  SR closed by agency-specific work management system          3.98
#> 9  SR closed by agency-specific work management system          0.91
#> 10 SR closed by agency-specific work management system          0.16
#> 12 SR closed by agency-specific work management system          3.31
#>                 address zip_code      neighborhood council_district
#> 1      733 LYNDHURST ST    21229 Edmondson Village                8
#> 2          710 KEVIN RD    21229    Rognel Heights                8
#> 3  1303 N WOODINGTON RD    21229 Edmondson Village                8
#> 4  1303 N WOODINGTON RD    21229 Edmondson Village                8
#> 5      4109 COLBORNE RD    21229 Edmondson Village                8
#> 6    1115 WILDWOOD PKWY    21229 Edmondson Village                8
#> 8   601 N WOODINGTON RD    21229 Edmondson Village                8
#> 9    3900 BLK STOKES DR    21229 Edmondson Village                8
#> 10    900 WILDWOOD PKWY    21229 Edmondson Village                8
#> 12      3800 HARLEM AVE    21229 Edmondson Village                8
#>    police_district latitude longitude                  geometry
#> 1     Southwestern 39.29630 -76.68110 POINT (427508.6 180947.3)
#> 2     Southwestern 39.29612 -76.68702 POINT (426998.3 180926.1)
#> 3     Southwestern 39.30233 -76.68711 POINT (426988.6 181615.3)
#> 4     Southwestern 39.30233 -76.68711 POINT (426988.6 181615.3)
#> 5     Southwestern 39.29760 -76.68469 POINT (427198.8 181090.8)
#> 6     Southwestern 39.29985 -76.68337 POINT (427311.3 181341.6)
#> 8     Southwestern 39.29415 -76.68569     POINT (427114 180708)
#> 9     Southwestern 39.30078 -76.68185 POINT (427442.8 181445.2)
#> 10    Southwestern 39.29787 -76.68419 POINT (427241.9 181121.3)
#> 12    Southwestern 39.29528 -76.68032 POINT (427576.8 180834.9)

# Get dirty alley service requests for multiple years using purrr::map_dfr()
purrr::map_dfr(
  c(2021, 2020),
  ~ get_area_requests(
    area = area,
    year = .x,
    request_type = "SW-Dirty Alley"
  )
)
#> ✔ Downloading "Customer_Service_Request311_2021_Present" from
#>   <https://egis.baltimorecity.gov/egis/rest/services/GeoSpatialized_Tables/ServiceRequest_311/FeatureServer/0>
#> Layer type: "Feature Layer"
#> Geometry type: "esriGeometryPoint"
#> Service CRS: "EPSG:2248"
#> Output CRS: "EPSG:2804"
#> Removing 2 duplicate 311 service requests.
#> ✔ Downloading "311_Customer_Service_Requests_2020.csv"
#> Layer type: "Table"
#> `x` has 0 rows with missing coordinates.
#> Removing 5 duplicate 311 service requests.
#> Simple feature collection with 171 features and 21 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 426955 ymin: 180668.8 xmax: 427620.5 ymax: 181472.1
#> Projected CRS: NAD83(HARN) / Maryland
#> First 10 features:
#>    service_request_num        sr_type method_received        created_date
#> 1          21-00053788 SW-Dirty Alley           Phone 2021-01-22 03:07:54
#> 2          21-00059271 SW-Dirty Alley           Phone 2021-01-24 11:54:01
#> 3          21-00051584 SW-Dirty Alley             API 2021-01-21 08:36:09
#> 4          21-00107202 SW-Dirty Alley           Phone 2021-02-12 03:57:40
#> 5          21-00110903 SW-Dirty Alley           Phone 2021-02-13 05:44:24
#> 6          21-00115345 SW-Dirty Alley             API 2021-02-15 06:46:25
#> 7          21-00142040 SW-Dirty Alley           Phone 2021-02-24 09:37:26
#> 8          21-00062237 SW-Dirty Alley           Phone 2021-01-25 10:55:46
#> 9          21-00139354 SW-Dirty Alley             API 2021-02-23 17:06:52
#> 10         21-00070593 SW-Dirty Alley          System 2021-01-28 08:57:46
#>    sr_status         status_date            due_date          close_date
#> 1     Closed 2021-01-29 10:01:00 2021-02-02 03:07:54 2021-01-29 10:01:21
#> 2     Closed 2021-02-03 09:12:00 2021-02-02 11:54:01 2021-02-03 09:12:54
#> 3     Closed 2021-01-22 09:34:00 2021-02-01 08:36:09 2021-01-22 09:34:27
#> 4     Closed 2021-02-23 09:24:00 2021-02-24 03:57:40 2021-02-23 09:25:02
#> 5     Closed 2021-02-16 10:24:00 2021-02-24 05:44:23 2021-02-16 10:24:58
#> 6     Closed 2021-02-23 09:26:00 2021-02-24 06:46:25 2021-02-23 09:27:20
#> 7     Closed 2021-03-04 10:18:00 2021-03-05 09:37:26 2021-03-04 10:18:37
#> 8     Closed 2021-01-25 10:55:00 2021-02-03 10:55:46 2021-01-25 10:56:15
#> 9     Closed 2021-03-03 09:21:00 2021-03-04 17:06:52 2021-03-03 09:21:41
#> 10    Closed 2021-02-04 10:51:00 2021-02-08 08:57:46 2021-02-04 10:51:52
#>         agency    last_activity  last_activity_date
#> 1  Solid Waste Service Response 2021-01-29 10:01:00
#> 2  Solid Waste Service Response 2021-02-03 09:12:00
#> 3  Solid Waste Service Response 2021-01-22 09:34:00
#> 4  Solid Waste Service Response 2021-02-23 09:24:00
#> 5  Solid Waste Service Response 2021-02-16 10:24:00
#> 6  Solid Waste Service Response 2021-02-23 09:26:00
#> 7  Solid Waste Service Response 2021-03-04 10:18:00
#> 8  Solid Waste Service Response 2021-01-25 10:55:00
#> 9  Solid Waste Service Response 2021-03-03 09:21:00
#> 10 Solid Waste Service Response 2021-02-04 10:51:00
#>                                           outcome days_to_close
#> 1                                  Work completed          7.29
#> 2                                  Work completed          9.89
#> 3                                  Work completed          1.04
#> 4  SR assessed and no cause for action determined         11.23
#> 5  SR assessed and no cause for action determined          3.19
#> 6  SR assessed and no cause for action determined          8.11
#> 7                                  Work completed          8.03
#> 8                                  Work completed          0.00
#> 9  SR assessed and no cause for action determined          7.68
#> 10                                 Work completed          7.08
#>                address zip_code      neighborhood council_district
#> 1    800 WILDWOOD PKWY    21229 Edmondson Village                8
#> 2  826 N WOODINGTON RD    21229    Rognel Heights                8
#> 3   905 MOUNT HOLLY ST    21229          Edgewood                8
#> 4     4000 COLBORNE RD    21229 Edmondson Village                8
#> 5       3801 ROKEBY RD    21229 Edmondson Village                8
#> 6     4009 COLBORNE RD    21229 Edmondson Village                8
#> 7  631 N WOODINGTON RD    21229 Edmondson Village                8
#> 8   704 MOUNT HOLLY ST    21229 Edmondson Village                8
#> 9      4010 GELSTON DR    21229 Edmondson Village                8
#> 10    3909 COLBORNE RD    21229 Edmondson Village                8
#>    police_district latitude longitude objectid                  geometry
#> 1     Southwestern 39.29641 -76.68416       NA POINT (427245.1 180958.8)
#> 2     Southwestern 39.29722 -76.68637       NA POINT (427053.6 181048.6)
#> 3     Southwestern 39.29815 -76.67980       NA POINT (427619.9 181153.5)
#> 4     Southwestern 39.29801 -76.68264       NA POINT (427375.7 181136.8)
#> 5     Southwestern 39.29923 -76.68060       NA POINT (427550.9 181273.1)
#> 6     Southwestern 39.29766 -76.68282       NA POINT (427359.6 181098.7)
#> 7     Southwestern 39.29505 -76.68574       NA POINT (427109.2 180807.5)
#> 8     Southwestern 39.29535 -76.67989       NA POINT (427613.6 180842.2)
#> 9     Southwestern 39.29579 -76.68288       NA POINT (427355.8 180890.3)
#> 10    Southwestern 39.29768 -76.68213       NA POINT (427419.8 181100.8)