Get streets

Get streets in Baltimore City by name with option to exclude streets by name, crop to a bounding box, or to filter to selected street types or functional classifications.

Usage

get_streets(
  street_name,
  exclude_name = NULL,
  street_type = NULL,
  sha_class = NULL,
  block_num = NULL,
  bbox = NULL,
  union = TRUE
)

Arguments

street_name

Street names to return. Required.

exclude_name

Street names to exclude

street_type

selected street subtypes to include. By default, the returned data includes all subtypes except alleys ("STRALY"). Options include c("STRALY", "STRPRD", "STRR", "STREX", "STRFIC", "STRNDR", "STRURD", "STCLN", "STRTN"). Not supported for

sha_class

selected SHA classifications to include. "all" selects all streets with an assigned SHA classification (around one-quarter of all street segments). Additional options include c("COLL", "LOC", "MART", "PART", "FWY", "INT")

block_num

Integer vector with block number, e.g. 300, or range of block numbers (e.g. c(100, 500)) to filter streets.

bbox

bbox to crop returned streets. Optional.

union

Logical. If TRUE, use st_union to combine geometry by fullname of the streets.

Value

OUTPUT_DESCRIPTION

Details

DETAILS

See also

Examples

get_streets(street_name = "UNIVERSITY PKWY")
#> Simple feature collection with 2 features and 1 field
#> Geometry type: MULTILINESTRING
#> Dimension:     XY
#> Bounding box:  xmin: 431490.7 ymin: 184537.1 xmax: 433526.9 ymax: 186030.6
#> Projected CRS: NAD83(HARN) / Maryland
#> # A tibble: 2 × 2
#>   fullname                                                              geometry
#>   <chr>                                                    <MULTILINESTRING [m]>
#> 1 E UNIVERSITY PKWY ((433012.2 184928.3, 433040.2 184906.5, 433058.1 184891.7, …
#> 2 W UNIVERSITY PKWY ((431490.7 186021.5, 431503.9 186030.6), (432938.4 185006.9…

get_streets(street_name = c("E FAYETTE", "ORLEANS"), block_num = c(1700, 3600))
#> Warning: attribute variables are assumed to be spatially constant throughout all geometries
#> Simple feature collection with 2 features and 1 field
#> Geometry type: MULTILINESTRING
#> Dimension:     XY
#> Bounding box:  xmin: 435026.7 ymin: 180717.9 xmax: 437393.1 ymax: 180941.1
#> Projected CRS: NAD83(HARN) / Maryland
#> # A tibble: 2 × 2
#>   fullname                                                              geometry
#>   <chr>                                                    <MULTILINESTRING [m]>
#> 1 E FAYETTE ST ((436092.7 180775.4, 436117.4 180777), (435315.1 180732.7, 43537…
#> 2 ORLEANS ST   ((435026.7 180867.6, 435088.5 180870.3, 435123.3 180872.4, 43513…