Add a layer to a gpplot2 map with area property categorized by type

Real property or parcel data is from the Maryland State Department of Assessment and Taxation and may include outdated or inaccurate information.

Usage

layer_area_property(
  area = NULL,
  bbox = NULL,
  data = NULL,
  type = c("improved", "vacant", "principal residence", "use", "building type", "value"),
  asis = FALSE,
  diag_ratio = NULL,
  dist = NULL,
  asp = NULL,
  crop = TRUE,
  trim = FALSE,
  show_area = FALSE,
  show_mask = FALSE,
  crs = pkgconfig::get_config("mapbaltimore.crs", 2804),
  ...
)

Arguments

area

sf object. If multiple areas are provided, they are unioned into a single sf object using sf::st_union()

bbox

bbox object defining area used to filter data. If an area is provided, the bounding box is ignored.

data

sf object including data in area

type

Real property variable to map. Options include c("improved", "vacant", "principal residence", "value"). Currently supports only one variable at a time.

asis

Logical. Default FALSE. If TRUE, use inherited data as is without cropping to area.

diag_ratio

ratio to set map extent based diagonal distance of area's bounding box. Ignored when dist is provided.

dist

buffer distance in meters. Optional.

asp

Aspect ratio of width to height as a numeric value (e.g. 0.33) or character (e.g. "1:3").

crop

If TRUE, data cropped to area or bounding box sf::st_crop() adjusted by the dist, diag_ratio, and asp parameters provided. Default TRUE.

trim

If TRUE, data trimmed to area with sf::st_intersection(). This option is not supported for any adjusted areas that use the dist, diag_ratio, or asp parameters. Default FALSE.

show_area

Logical. Default FALSE. If TRUE, add an outline of the area to the layer.

show_mask

Logical. Default FALSE. If TRUE, add a mask using layer_area_mask

crs

Coordinate Reference System (CRS) to use for the returned data. The CRS of the provided data and bounding box or area must match one another but are not required to match the CRS provided by this parameter.

...

passed to ggplot2::geom_sf() for data layer.

See also

Examples

if (FALSE) {
area <- get_area("neighborhood", "West Forest Park")

property <- get_area_property(area = area)

ggplot2::ggplot() +
  layer_area_property(area = area, data = property, type = "principal residence")
}